Compiler on top of Java for special lexical specifications, all documented.
Designed and implemented a scanner for a programming language whose
lexical specifications are given below. The scanner identifies and
outputs tokens (valid words and punctuation) in the source program.

Its output is a token that can thereafter be used by the syntactic
analyzer to verify that the program is
syntactically valid. When called, the lexical analyzer should extract
the next token from the source
program. The lexical analyzer should be able to output a token even if
the input does not form a correct program.

The syntax of the language will is specified in the docs.

Appropriate test cases are provided that test for a wide variety of
valid and invalid cases.